requirements.txt
qdrant-client==1.7.0
openai==1.3.5
loguru==0.7.2
python-dotenv==0.13.0
jupyterlab==4.0.10
pandas==2.0.3
sentence-transformers==2.2.2
httpx==0.24.1
streamlit>=1.28.0
plotly>=5.17.0             
 streamlit-option-menu>=0.3.6 
streamlit-aggrid>=0.3.4     
wordcloud>=1.9.2           
jieba>=0.42.1             
seaborn>=0.12.0            